+++ pages/dappstore/src/pages/landing/partials/ecosystem-card-grid.tsx
@@ -2,19 +2,20 @@
 // SPDX-License-Identifier:ENCL-1.0(https://github.com/evmos/apps/blob/main/LICENSE)
 
 import { cn } from "helpers";
+import React from "react";
 import { ComponentProps } from "react";
 
 export const EcosystemCardGrid = ({
   className,
   ...rest
-}: ComponentProps<"section">) => {
+}: ComponentProps<"section"> & { children: React.ReactNode }) => {
   return (
     <section
       className={cn(
-        "grid gap-8 sm:grid-cols-2 lg:grid-cols-3 xl:grid-cols-4",
+        "grid gap-4 grid-cols-2 md:gap-8 lg:grid-cols-4 mb-10",
         className,
       )}
       {...rest}
-    />
+    ></section>
   );
 };