+++ CHANGELOG.md
@@ -36,7 +36,22 @@ Ref: https://keepachangelog.com/en/1.0.0/
 
 ## Unreleased
 
+-(fix) DAPP-14 | apps/mission 1.0.37 packages/i18n 1.0.6 packages/stateful-components 1.0.7 packages/ui-helpers 1.0.24 packages/ui 1.0.2 pages/dappstore 1.1.10 | dapps detail pages, airfoil design
 -(fix) DAPP-22 | packages/ui 1.0.1 | storybooks: chips
+
+## 2.0.14 - 2024-04-25
+
+- (feat) apps/mission 1.0.36 | Implements script to sync notion images with vercel blobs
+
+## 2.0.13 - 2024-04-12
+
+- (feat) pages/governance 1.1.4 | Read blocked proposals from notion
+- (fix) dapp-58 | pages/dappstore | Update ecosystem section - show 8 dapps
+
+## 2.0.12 - 2024-03-18
+
+- (fix) xap-151 | packages/widgets 1.0.8 packages/instant-dapps 1.0.5 pages/dapp-store 1.1.8 | Add Leap Elements instant dapp
+- (fix) fse-975 | packages/trpc 1.0.3 | Fix format address for staked endpoint
 
 ## 2.0.11 - 2024-03-13
 
@@ -55,6 +70,7 @@ Ref: https://keepachangelog.com/en/1.0.0/
 
 - (chore) | packages/icons 1.0.13 | Fix icons for wallets
 - (chore) | apps/mission 1.0.33 | Swap rpcs by the newly crated proxy
+
 - (chore) | packages/evmos-wallet 1.0.31 | Fix provider request
 - (chore) onb-4 | apps/mission 1.0.33 packages/constants-helpers 1.0.12 packages/config 1.0.2 packages/evmos-wallet 1.0.30 packages/helpers 1.0.13 packages/stateful-components 1.0.6 packages/trpc 1.0.2 packages/ui-helpers 1.0.23 pages/dappstore 1.1.6 pages/portfolio 1.1.11 | Add Leap Wallet
 - (chore) onb-3 | packages/evmos-wallet 1.0.29 packages/stateful-components 1.0.5 pages/staking 1.1.4 pages/vesting 1.0.2 | Add Rabby Wallet